Add files via upload
Objective: The primary objective is to retrieve the retail sales dataset using the Kaggle API and then preprocess the data to ensure it is clean and ready for analysis. Data preprocessing is a critical first step in any data analysis project, as it helps identify issues with the raw dataset, such as missing values, incorrect data types, or duplicates. In this part of the project, we also perform exploratory data analysis (EDA) to understand the structure of the dataset and gain initial insights